+++ app/Http/Controllers/BlogsController.php
@@ -417,13 +417,15 @@ class BlogsController extends Controller
             return back();
         }
 
-        // update the likes
-        --$blog->likes;
-        $blog->save();
-
         // delete the like
         $response = Like::where('object_id', '=', $id)->where('user_id', '=', $this->user->id)->where('object_type', '=', 'blog')->first();
-        $response->delete();
+        if ($response) {
+            $response->delete();
+
+            // update the likes
+            --$blog->likes;
+            $blog->save();
+        }
 
         flash()->success('Success', 'You are no longer liking the blog.');
